+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/AbstractBaseScreen.java
@@ -138,7 +138,8 @@ public abstract class AbstractBaseScreen<T extends AbstractContainerMenu> extend
             mouseX,
             mouseY
         );
-        titleMarquee.render(graphics, titleLabelX, titleLabelY, font, hoveringOverTitle);
+        titleMarquee.render(graphics, titleLabelX, titleLabelY, font, hoveringOverTitle,
+            minecraft.getDeltaTracker().getGameTimeDeltaTicks());
         renderPlayerInventoryTitle(graphics);
     }