+++ CHANGELOG.md
@@ -16,6 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   The mouse cursor can now go out of the scrollbar bounds while scrolling.
 -   You can no longer change the energy capacity to be zero.
 -   Fixed stonecutter patterns not respecting output larger than one.
+-   Fixed title marquee effect when hovering being FPS dependant.
 
 ## [3.0.0-beta.3] - 2026-04-16